requirements.txt
gradio
torch
transformers
opencv-python-headless
Pillow
